+++ lib/loan_creator/bullet.rb
@@ -7,7 +7,7 @@ module LoanCreator
       reset_current_term
       @crd_beginning_of_period = amount
       @crd_end_of_period = amount
-      (duration_in_periods - 1).times { |period| compute_term(timetable, period + 1) }
+      (duration_in_periods - 1).times { |period| compute_term(timetable) }
       compute_last_term
       timetable << current_term
       timetable
@@ -17,22 +17,22 @@ module LoanCreator
 
     def compute_last_term
       @crd_end_of_period                         = bigd('0')
-      @period_interests                          = compute_capitalized_interests(duration_in_periods)
+      @capitalized_interests_beginning_of_period = @capitalized_interests_end_of_period
+      @period_interests                          = @capitalized_interests_end_of_period + compute_capitalized_interests
+      @capitalized_interests_end_of_period       = 0
       @period_capital                            = @crd_beginning_of_period
       @total_paid_capital_end_of_period          = @period_capital
       @total_paid_interests_end_of_period        = @period_interests
       @period_amount_to_pay                      = @period_capital + @period_interests
-      @capitalized_interests_beginning_of_period = @period_interests
-      @capitalized_interests_end_of_period       = 0
     end
 
-    def compute_capitalized_interests(period)
-      amount.mult((bigd(1) + periodic_interests_rate) ** period, BIG_DECIMAL_DIGITS) - amount
+    def compute_capitalized_interests
+      (amount + @capitalized_interests_beginning_of_period).mult(periodic_interests_rate, BIG_DECIMAL_DIGITS)
     end
 
-    def compute_term(timetable, period)
-      @capitalized_interests_beginning_of_period = compute_capitalized_interests(period)
-      @capitalized_interests_end_of_period = @capitalized_interests_beginning_of_period
+    def compute_term(timetable)
+      @capitalized_interests_beginning_of_period = @capitalized_interests_end_of_period
+      @capitalized_interests_end_of_period += compute_capitalized_interests
       timetable << current_term
     end
   end