+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.7.0
 
 - change `capitalized_interests` for `capitalized_interests_beginning_of_period`
   and `capitalized_interests_end_of_period` in `LoanCreator::Term`
+- add `capitalized_interests` in `:inital_values` for `LoanCreator::Bullet` loans
 
 v0.6.2
 -------------------------